refactor: Move real LN fixtures into fedimint-testing
