+++ fedimint-testing/src/ln/mock.rs
